feat: Moved structure data classes to `rci` folder
